% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_categorias.R
\name{plot_bar_categorias}
\alias{plot_bar_categorias}
\title{Barplot horizontal de categorias}
\usage{
plot_bar_categorias(df, coluna, fill_colors = NULL)
}
\arguments{
\item{df}{Data frame contendo a variável categórica.}

\item{coluna}{Nome da coluna (sem aspas, tidy evaluation).}

\item{fill_colors}{Vetor de cores nomeadas (opcional).}
}
\value{
Um gráfico ggplot2.
}
\description{
Barplot horizontal de categorias
}
\examples{
df <- data.frame(grupo = c("A","B","A","C"))
plot_bar_categorias(df, coluna = grupo)
}
