% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot_enriquecimento.R
\name{plot_dotplot_enriquecimento}
\alias{plot_dotplot_enriquecimento}
\title{Dotplot de enriquecimento funcional}
\usage{
plot_dotplot_enriquecimento(df, via, log10_col, genes_col)
}
\arguments{
\item{df}{Data frame contendo vias, log10(FDR) e número de genes.}

\item{via}{Coluna com nome da via.}

\item{log10_col}{Coluna com valor -log10(FDR).}

\item{genes_col}{Coluna com número de genes.}
}
\value{
Um gráfico ggplot2.
}
\description{
Dotplot de enriquecimento funcional
}
