% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset_plot.R
\name{plot_enriquecimento}
\alias{plot_enriquecimento}
\title{Dotplot de enriquecimento funcional}
\usage{
plot_enriquecimento(df)
}
\arguments{
\item{df}{Data frame com colunas: Via, Database, Log10_FDR, Input_number.}
}
\value{
Um grafico ggplot2.
}
\description{
Gera um dotplot mostrando Log10_FDR vs vias, com tamanho proporcional ao numero de genes
e forma de acordo com o banco de dados de referencia.
}
