% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Heatmap simples}
\usage{
plot_heatmap(mat, scale = FALSE)
}
\arguments{
\item{mat}{Matriz numérica.}

\item{scale}{Se TRUE, aplica padronização por linha.}
}
\value{
Um heatmap gerado por pheatmap.
}
\description{
Heatmap simples
}
