% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_chart.R
\name{plot_pie}
\alias{plot_pie}
\title{Gráfico de pizza simples}
\usage{
plot_pie(valores, categorias, title = "")
}
\arguments{
\item{valores}{Vetor numérico com os valores.}

\item{categorias}{Vetor de nomes das categorias.}

\item{title}{Título do gráfico.}
}
\value{
Um gráfico de pizza (base R)
}
\description{
Gráfico de pizza simples
}
